Move markdown KBA support into separate package including readme
