+++ src/main/java/sirius/biz/tycho/kb/markdown/KnowledgeBaseMarkdownDocument.java
@@ -6,7 +6,7 @@
  * https://www.scireum.de - info@scireum.de
  */
 
-package sirius.biz.tycho.kb;
+package sirius.biz.tycho.kb.markdown;
 
 import sirius.kernel.commons.Strings;
